+++ src/main/scala/example/hawaii/package.scala
@@ -1,0 +1,4 @@
+package example
+
+package object hawaii {
+}
